👌 IMPROVE: move dataclasses into data package to unify with data/repository
