+++ app/src/main/java/com/github/swent/echo/data/model/Association.kt
@@ -1,3 +1,3 @@
-package com.github.swent.echo.classes
+package com.github.swent.echo.data.model
 
 data class Association(val associationId: String, val name: String, val description: String)
