+++ app/src/main/java/com/github/swent/echo/data/model/Tag.kt
@@ -1,3 +1,3 @@
-package com.github.swent.echo.classes
+package com.github.swent.echo.data.model
 
 data class Tag(val tagId: String, val name: String)
