+++ app/src/main/java/com/github/swent/echo/data/model/User.kt
@@ -1,3 +1,3 @@
-package com.github.swent.echo.classes
+package com.github.swent.echo.data.model
 
 data class User(val userId: String, val userProfileId: UserProfile)
