+++ app/src/main/java/com/github/swent/echo/data/model/UserProfile.kt
@@ -1,3 +1,3 @@
-package com.github.swent.echo.classes
+package com.github.swent.echo.data.model
 
 data class UserProfile(val userId: String, val name: String)
